# api/models.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"time"
)

// Defines values for Status.
const (
	Done    Status = "done"
	Waiting Status = "waiting"
	Working Status = "working"
)

// Status The task status
type Status string

// Task defines model for Task.
type Task struct {
	// CreatedAt The task creation date
	CreatedAt time.Time `json:"createdAt"`

	// Description The task description
	Description *string `json:"description,omitempty"`

	// Id The task identifier
	Id uint64 `json:"id"`

	// Status The task status
	Status Status `json:"status"`

	// Title The task title
	Title string `json:"title"`
}

// TaskData defines model for TaskData.
type TaskData struct {
	// Description The task description
	Description *string `json:"description,omitempty"`

	// Status The task status
	Status Status `json:"status"`

	// Title The task title
	Title string `json:"title"`
}

// ErrorResponse defines model for errorResponse.
type ErrorResponse struct {
	// Details Error details
	Details string `json:"details"`
}

// ListTasksParams defines parameters for ListTasks.
type ListTasksParams struct {
	// Status Filters the tasks by their status
	Status *Status `form:"status,omitempty" json:"status,omitempty"`
}

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = TaskData

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody = TaskData